requirements.txt
anyio==4.7.0
asgiref==3.8.1
attrs==24.2.0
boto3==1.35.76
botocore==1.35.76
certifi==2024.8.30
cffi==1.17.1
click==8.1.7
cryptography==44.0.0
Django==5.1.3
django-cors-headers==4.6.0
django-extensions==3.2.3
django-rq==3.0.0
django-storages==1.14.4
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
drf-spectacular==0.28.0
gunicorn==23.0.0
h11==0.14.0
httpcore==1.0.7
httpx==0.28.1
idna==3.10
inflection==0.5.1
iniconfig==2.0.0
jmespath==1.0.1
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
packaging==24.2
pillow==11.0.0
pluggy==1.5.0
psycopg2-binary==2.9.10
pycparser==2.22
PyJWT==2.10.1
pytest==8.3.4
pytest-django==4.9.0
python-dateutil==2.9.0.post0
PyYAML==6.0.2
redis==5.2.1
referencing==0.35.1
rpds-py==0.22.3
rq==2.0.0
s3transfer==0.10.4
six==1.17.0
sniffio==1.3.1
sqlparse==0.5.2
toml==0.10.2
typing_extensions==4.12.2
uritemplate==4.1.1
urllib3==2.2.3
